Fix api url
