+++ next.config.js
@@ -7,7 +7,7 @@ const nextConfig = {
         destination:
           process.env.NODE_ENV === 'development'
             ? 'http://127.0.0.1:5328/api/:path*'
-            : '/api/',
+            : 'https://drone.ziasvannes.tech/api/:path*',
       },
     ]
   },
